Fix matc test cases
